fix : directory constructure
